softwares/sistema_de_autoavaliacao_de_seguranca_cibernetica/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Mail
Flask-Bcrypt
Flask-Migrate
email-validator
psycopg2-binary
Pillow
PyJWT
gunicorn
python-dotenv
matplotlib
pandas
plotly
dash
pyinstaller